refactor: move chain definitions to config/chains
