+++ config/chains.ts
@@ -1,0 +1,10 @@
+export enum CHAINS {
+  Mainnet = 1,
+  Holesky = 17000,
+  Hoodi = 560048,
+  Sepolia = 11155111,
+  Optimism = 10,
+  OptimismSepolia = 11155420,
+  Unichain = 130,
+  UnichainSepolia = 1301,
+}
