+++ config/index.ts
@@ -2,4 +2,5 @@ export * from './get-config';
 export * from './get-secret-config';
 export * from './provider';
 export * from './use-config';
+export * from './chains';
 export * from './known-chain-ids';
