+++ config/known-chain-ids.ts
@@ -1,12 +1,5 @@
-// Chain IDs explicitly supported by the widget UI.
-// Must stay in sync with the CHAINS enum in consts/chains.ts.
-export const KNOWN_CHAIN_IDS = new Set([
-  1, // Mainnet
-  17000, // Holesky
-  560048, // Hoodi
-  11155111, // Sepolia
-  10, // Optimism
-  11155420, // OptimismSepolia
-  130, // Unichain
-  1301, // UnichainSepolia
-]);
+import { CHAINS } from './chains';
+
+export const KNOWN_CHAIN_IDS = new Set<number>(
+  Object.values(CHAINS).filter((v): v is number => typeof v === 'number'),
+);
